pyproject.toml
[project]
name = "code-review"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "fastapi>=0.115.14",
    "flower>=2.0.1",
    "openai>=1.93.0",
    "pygithub>=2.6.1",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]
